+++ src/main.ts
@@ -4,6 +4,7 @@ import "@fontsource/outfit/400.css";
 import "@fontsource/outfit/500.css";
 import "@fontsource/outfit/600.css";
 import "@fontsource/outfit/700.css";
+import "@fontsource/material-symbols-outlined/400.css";
 import "./style.css";
 import App from "./App.vue";
 import router from "./router";